requirements.txt
requests==2.31.0
python-dateutil==2.8.2
typing-extensions==4.9.0
flask==3.0.2
flask-wtf==1.2.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
textstat==0.7.3 